Initial push. Minor change. Moving BinaryToolsTest to correct and respective package. This did not get moved upon original Tools transfer.
