+++ test/smelter/filesmelter/OHLCV_BinaryLexicalFileUnsmelterTest.java
@@ -1,0 +1,16 @@
+/**
+ * @author Bruce Lamb
+ * @since 17 JUN 2025
+ */
+package tradedatacorp.smelter.filesmelter;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class OHLCV_BinaryLexicalFileUnsmelterTest{}
