+++ test/tools/binarytools/BinaryToolsTest.java
@@ -2,7 +2,7 @@
  * @author Bruce Lamb
  * @since 20 JAN 2025
  */
-package tradedatacorp.smelter.lexical.binary;
+package tradedatacorp.tools.binaryTools;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
